draft changes for combining gtest binaries
