+++ test/gtest/conv_f8_wrw.cpp
@@ -33,6 +33,7 @@
 #include "get_handle.hpp"
 #include "f8_cast_util.hpp"
 
+template<>
 std::vector<Conv3DTestCase> ConvTestConfigs()
 { // g   n   c   d    h   w   k   z  y  x pad_x pad_y pad_z stri_x stri_y stri_z dia_x dia_y dia_z
     return {{1, 16, 16, 1, 14, 14, 16, 1, 3, 3, 1, 1, 0, 1, 1, 1, 1, 1, 1, miopenConvolution},
@@ -195,5 +196,5 @@ INSTANTIATE_TEST_SUITE_P(
     ConvWrwTest,
     ConvWrwSolverTestF8,
     testing::Combine(testing::Values(miopenConvolutionBwdWeightsAlgoImplicitGEMM),
-                     testing::ValuesIn(ConvTestConfigs()),
+                     testing::ValuesIn(ConvTestConfigs<Conv3DTestCase>()),
                      testing::Values(miopenTensorNDHWC)));
