+++ test/gtest/group_conv3d_wrw.hpp
@@ -27,6 +27,7 @@
 
 #include "conv3d_test_case.hpp"
 
+template<>
 std::vector<Conv3DTestCase> ConvTestConfigs()
 { // g   n   c   d    h   w   k   z  y  x pad_x pad_y pad_z stri_x stri_y stri_z dia_x dia_y dia_z
     return {{1, 128, 64, 14, 28, 28, 64, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, miopenConvolution},
