+++ test/gtest/group_solver.hpp
@@ -32,29 +32,9 @@
 
 #include "../driver/tensor_driver.hpp"
 #include "conv_common.hpp"
+#include "conv_test_base.hpp"
 
-template <typename T>
-miopenDataType_t GetDataType();
-
-template <>
-miopenDataType_t GetDataType<float>()
-{
-    return miopenFloat;
-}
-
-template <>
-miopenDataType_t GetDataType<half_float::half>()
-{
-    return miopenHalf;
-}
-
-template <>
-miopenDataType_t GetDataType<int8_t>()
-{
-    return miopenInt8;
-}
-
-struct ConvTestCase
+struct ConvTestCaseGroup
 {
     size_t G;
     size_t N;
@@ -71,7 +51,7 @@ struct ConvTestCase
     size_t dilation_x;
     size_t dilation_y;
     miopenConvolutionMode_t conv_mode;
-    friend std::ostream& operator<<(std::ostream& os, const ConvTestCase& tc)
+    friend std::ostream& operator<<(std::ostream& os, const ConvTestCaseGroup& tc)
     {
         return os << " G:" << tc.G << " N:" << tc.N << " C:" << tc.C << " H:" << tc.H
                   << " W:" << tc.W << " k:" << tc.k << " y:" << tc.y << " x:" << tc.x
@@ -101,16 +81,6 @@ struct ConvTestCase
     }
 };
 
-std::vector<ConvTestCase> ConvTestConfigs()
-{ // g  n  c   h   w   k   y  x pad_x pad_y stri_x stri_y dia_x dia_y
-    return {{1, 256, 192, 28, 28, 192, 3, 3, 1, 1, 1, 1, 1, 1, miopenConvolution},
-            {1, 256, 12, 28, 28, 12, 3, 3, 1, 1, 1, 1, 1, 1, miopenConvolution},
-            {4, 256, 192, 28, 28, 192, 3, 3, 1, 1, 1, 1, 1, 1, miopenConvolution},
-            {8, 256, 192, 28, 28, 192, 3, 3, 1, 1, 1, 1, 1, 1, miopenConvolution},
-            {8, 256, 384, 28, 28, 384, 3, 3, 1, 1, 1, 1, 1, 1, miopenConvolution},
-            {32, 256, 1024, 28, 28, 2048, 3, 3, 1, 1, 1, 1, 1, 1, miopenConvolution}};
-}
-
 inline int SetTensorLayout(miopen::TensorDescriptor& desc)
 {
     // get layout string names
@@ -124,9 +94,9 @@ inline int SetTensorLayout(miopen::TensorDescriptor& desc)
 }
 
 template <typename T = float>
-struct ConvFwdSolverTest
+struct ConvGroupFwdSolverTest
     : public ::testing::TestWithParam<
-          std::tuple<miopenConvFwdAlgorithm_t, ConvTestCase, miopenTensorLayout_t>>
+          std::tuple<miopenConvFwdAlgorithm_t, ConvTestCaseGroup, miopenTensorLayout_t>>
 {
 protected:
     void SetUp() override
@@ -183,7 +153,7 @@ protected:
         EXPECT_TRUE(error < threshold)
             << "Error beyond tolerance Error:" << error << ",  Threshold: " << threshold;
     }
-    ConvTestCase conv_config;
+    ConvTestCaseGroup conv_config;
     miopen::ConvolutionDescriptor conv_desc;
     tensor<T> input;
     tensor<T> weights;
