+++ test/gtest/nonpack_conv3d_fwd.cpp
@@ -94,5 +94,5 @@ TEST_P(ConvFwdSolverTest3D, CKNonPackConvFwd3D)
 INSTANTIATE_TEST_SUITE_P(ConvFwdTest,
                          ConvFwdSolverTest3D,
                          testing::Combine(testing::Values(miopenConvolutionFwdAlgoImplicitGEMM),
-                                          testing::ValuesIn(ConvTestConfigs()),
+                                          testing::ValuesIn(ConvTestConfigs<NonPackTestCase>()),
                                           testing::Values(miopenTensorNDHWC)));
