+++ test/gtest/solver_bwd_f8.cpp
@@ -60,10 +60,10 @@ TEST_P(ConvBwdFp8Naive, DISABLED_Bwd)
 INSTANTIATE_TEST_SUITE_P(ConvBwdTest,
                          ConvBwdFp8,
                          testing::Combine(testing::Values(miopenConvolutionAlgoGEMM),
-                                          testing::ValuesIn(GetNetwork1())));
+                                          testing::ValuesIn(GetNetwork1<ConvTestCaseBase>())));
 // Since NaiveConv is verified against the CPU, we are conservative in the number and type
 // of test cases we instantiate
 INSTANTIATE_TEST_SUITE_P(ConvBwdTest,
                          ConvBwdFp8Naive,
                          testing::Combine(testing::Values(miopenConvolutionAlgoGEMM),
-                                          testing::ValuesIn(ConvTestConfigs())));
+                                          testing::ValuesIn(ConvTestConfigs<ConvTestCaseBase>())));
