+++ test/gtest/solver_fwd_f8.cpp
@@ -60,12 +60,12 @@ TEST_P(ConvFwdFp8Naive, DISABLED_Fwd)
 INSTANTIATE_TEST_SUITE_P(ConvFwdTest,
                          ConvFwdFp8,
                          testing::Combine(testing::Values(miopenConvolutionAlgoGEMM),
-                                          testing::ValuesIn(ConvTestConfigs()),
+                                          testing::ValuesIn(ConvTestConfigs<ConvTestCaseBase>()),
                                           testing::Values(miopenTensorNCHW)));
 // Since NaiveConv is verified against the CPU, we are conservative in the number and type
 // of test cases we instantiate
 INSTANTIATE_TEST_SUITE_P(ConvFwdTest,
                          ConvFwdFp8Naive,
                          testing::Combine(testing::Values(miopenConvolutionAlgoGEMM),
-                                          testing::ValuesIn(ConvTestConfigs()),
+                                          testing::ValuesIn(ConvTestConfigs<ConvTestCaseBase>()),
                                           testing::Values(miopenTensorNCHW)));
